% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{check.db_conn}
\alias{check.db_conn}
\title{Check a Database Connection (OBSOLETE)}
\usage{
check.db_conn(this.conn, pass, ...)
}
\arguments{
\item{this.conn}{A \code{\link[DBI]{DBI-package}} connection object}

\item{pass}{The password to use if required: raw input is converted to character}

\item{...}{Not used}
}
\value{
An active DBI connection object
}
\description{
\code{check.db_conn} Validates a DBI connection and reconnects if it is invalid by using the connection's stored settings.
}
\seealso{
Other Data transmission: 
\code{\link{export_data}()},
\code{\link{get_data}()},
\code{\link{make_query}()}
}
\concept{Data transmission}
