% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_integrity.R
\name{check.env}
\alias{check.env}
\title{Environment Integrity Check}
\usage{
check.env(...)
}
\arguments{
\item{...}{One or more environments to check.}
}
\value{
A logical scalar indicating success or failure. In the case of
failure (\code{FALSE}), missing objects are attached in attribute
\code{"missing"}.
}
\description{
\code{check.env()} checks one or more environments for required objects.
}
\section{Changelog}{

\describe{
\item{Version 0.1.2.1000}{
\itemize{
\item{Changed the output to a conditionally-attributed logical scalar.}
}
}
}
}

\seealso{
Other Chapter 1 - Environment Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%-=\%}()},
\code{\link{\%must.have\%}()}
}
\concept{Chapter 1 - Environment Integrity}
