% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{do.export_data}
\alias{do.export_data}
\title{Export Tabular Data From R to DBMS}
\usage{
do.export_data(
  out.data,
  this.conn,
  tbl = NULL,
  sch = "dbo",
  post.cmd = NULL,
  append = FALSE,
  persist.conn = TRUE,
  ...
)
}
\arguments{
\item{out.data}{The data to write to the database, preferably a data.frame or coercible}

\item{this.conn}{A \code{\link[DBI]{DBI-package}} connection object}

\item{tbl}{(string) The name of the table to drop (if it exists) and create}

\item{sch}{(string) The name of the schema in which to create the table}

\item{post.cmd}{(string) When provided, a post-table creation call to sqlQuery() will be executed with `post.cmd` as the query string: use this for subsequent operations such as indexing, execution of additional stored procedures, etc.}

\item{append}{(logical)    Should the data be appended to existing data rather than replacing existing data?}

\item{persist.conn}{(logical | TRUE) When \code{FALSE}, the connection will be closed at the end of execution}

\item{...}{Additional arguments passed to \code{\link[DBI]{dbWriteTable}}}
}
\description{
\code{do.export_data} serves as a wrapper for \code{\link[DBI]{dbWriteTable}}. When using \code{append=FALSE}, the operation is destructive as the target table is dropped if it exists, so use this function with caution.
The parameter `post.cmd` is provided
}
\seealso{
Other Data transmission: 
\code{\link{check.db_conn}()},
\code{\link{do.get_data}()},
\code{\link{do.make_query}()},
\code{\link{do.make_workers}()}
}
\concept{Data transmission}
