% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{do.get_data}
\alias{do.get_data}
\title{Retrieve SQL Data and Post-process}
\usage{
do.get_data(
  this.conn,
  src.name = NULL,
  tgt.name = NULL,
  this.data = NULL,
  post.op = eval,
  verbose = FALSE,
  promise = FALSE,
  persist.conn = TRUE,
  ...
)
}
\arguments{
\item{this.conn}{A \code{\link[DBI]{DBI-package}} connection object}

\item{src.name}{(string) Data source name}

\item{tgt.name}{(string | NULL) When not \code{NULL}, the name of the R workspace object to use when assigning the output. Use format \code{obj_name@env_name} to specify a target environment to assign the object}

\item{this.data}{(expression) When not \code{NULL}, an R object for post-operation only}

\item{post.op}{(object) A function or list of functions to process the retrieved dataset (argument #1) before assigning to workspace}

\item{verbose}{(logical | FALSE) When \code{TRUE}, interim output is printed to console}

\item{promise}{(logical | FALSE) When \code{TRUE}, assignment is done via a call to \code{\link[future]{futureAssign}}}

\item{persist.conn}{(logical | TRUE) When \code{FALSE}, the connection will be closed at the end of execution}

\item{...}{Arguments to be used by function \code{\link{do.make_query}}}
}
\value{
Invisibly, the retrieved, post-processed dataset
}
\description{
The workflow for this function is to retrieve the dataset from a source, post-process it according to the value of argument \code{post.op}, and, when \code{tgt.name} is not \code{NULL}, assign it to an object named according to the value of \code{tgt.name}
}
\seealso{
Other Data transmission: 
\code{\link{check.db_conn}()},
\code{\link{do.export_data}()},
\code{\link{do.make_query}()},
\code{\link{do.make_workers}()}
}
\concept{Data transmission}
