% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_environment_processing.R
\name{do.get_pkgs}
\alias{do.get_pkgs}
\title{Install and load packages}
\usage{
do.get_pkgs()
}
\description{
`do.get_pkgs()` is no longer in use
}
\seealso{
Other DEPRECATED: 
\code{\link{do.make_cluster}()},
\code{\link{terminate.cluster_workers}()}
}
\concept{DEPRECATED}
