% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{do.make_cluster}
\alias{do.make_cluster}
\title{A Wrapper for \code{\link[parallelly]{makeClusterPSOCK}}}
\usage{
do.make_cluster(...)
}
\description{
`do.make_cluster()` is no longer in use
}
\seealso{
Other DEPRECATED: 
\code{\link{do.get_pkgs}()},
\code{\link{terminate.cluster_workers}()}
}
\concept{DEPRECATED}
