% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_workflow_management.R
\name{execute.workflow}
\alias{execute.workflow}
\title{Execute a Stored Workflow Step}
\usage{
execute.workflow(wf, wf_step, list.only = FALSE)
}
\arguments{
\item{wf}{(list) The workflow queue object containing the quoted workflow steps}

\item{wf_step}{(string/symbol) The names of the steps to execute.  Tip: label the steps in the order they should execute.}

\item{list.only}{(logical | FALSE) When \code{TRUE}, available workflows and workflow steps are printed to console before the function exits.}
}
\value{
When \code{list.only == TRUE}, the listing of workflows and corresponding steps invisibly; otherwise, nothing.
}
\description{
\code{execute.workflow} executes quoted expressions referencing \code{read.snippet} calls.
}
