% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{export_data}
\alias{export_data}
\title{Export Tabular Data From R to DBMS}
\usage{
export_data(
  this.conn,
  out.data,
  tbl = NULL,
  sch = "dbo",
  append = FALSE,
  persist.conn = TRUE,
  ...
)
}
\arguments{
\item{this.conn}{A \code{\link[DBI]{DBI-package}} connection object}

\item{out.data}{The data to write to the database, preferably a data.frame or coercible}

\item{tbl}{(string) The name of the table to drop (if it exists) and create}

\item{sch}{(string) The name of the schema in which to create the table}

\item{append}{(logical)    Should the data be appended to existing data rather than replacing existing data?}

\item{persist.conn}{(logical | \code{TRUE}) When \code{FALSE}, the connection will be closed at the end of execution}

\item{...}{Additional arguments passed to \code{\link[DBI]{dbWriteTable}}}
}
\description{
\code{export_data} serves as a wrapper for \code{\link[DBI]{dbWriteTable}} and \code{\link[DBI]{dbAppendTable}}. When using \code{append=FALSE}, the operation is destructive as the target table is dropped if it exists, so use this function with caution.
}
\seealso{
Other Data transmission: 
\code{\link{check.db_conn}()},
\code{\link{get_data}()},
\code{\link{make_query}()}
}
\concept{Data transmission}
