% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{get.cluster_meta}
\alias{get.cluster_meta}
\title{Retrieve Cluster Metadata}
\usage{
get.cluster_meta(cl)
}
\arguments{
\item{cl}{A cluster object}
}
\value{
Invisibly, an environment with active-binding \code{cluster_meta}
}
\description{
\code{get.cluster_meta} reads the information from a saved object created with one of the \code{makeCluster} family of functions
}
\concept{Parallelism}
