% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{get_data}
\alias{get_data}
\title{Retrieve T-SQL Data}
\usage{
get_data(
  this.conn,
  src.name = NULL,
  tgt.name = NULL,
  this.data = NULL,
  post.op = eval,
  chatty = FALSE,
  promise = FALSE,
  persist.conn = TRUE,
  ...
)
}
\arguments{
\item{this.conn}{A \code{\link[DBI]{DBI-package}} connection object}

\item{src.name}{(string) Data source name}

\item{tgt.name}{(string | \code{NULL}) When not \code{NULL}, the name of the R workspace object to use when assigning the output. Use format \code{obj_name@env_name} to specify a target environment to assign the object}

\item{this.data}{(expression) When not \code{NULL}, an R object for post-operation only}

\item{post.op}{(object) A function or list of functions to process the retrieved dataset (argument #2) before assigning to workspace}

\item{chatty}{(logical | \code{FALSE}) Verbosity flag}

\item{promise}{(logical | \code{FALSE}) When \code{TRUE}, assignment is done via a call to \code{\link[future]{futureAssign}}}

\item{persist.conn}{(logical | \code{TRUE}) When \code{FALSE}, the connection will be closed at the end of execution}

\item{...}{Arguments to be used by function \code{\link{make_query}}}
}
\value{
Invisibly, the retrieved, post-processed dataset
}
\description{
The workflow for this function is to retrieve the data from SQL Server, post-process it according to the value of argument \code{post.op}, and, when \code{tgt.name} is not \code{NULL}, assign it to an object named according to the value of \code{tgt.name}
}
\seealso{
Other Data transmission: 
\code{\link{export_data}()},
\code{\link{make_query}()}
}
\concept{Data transmission}
