% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_environment_integrity.R
\name{\%check\%}
\alias{\%check\%}
\title{Environment Integrity Check}
\usage{
env \%check\% x
}
\arguments{
\item{env}{(object | string) An environment or name of an environment}

\item{x}{(string[]) The names of the objects that \code{env} must have}
}
\value{
If all objects that \code{env} \code{\link{must.have}} are present, nothing is returned; otherwise, an error message is returned.
}
\description{
\code{\%check\%} checks the environment for required objects
}
\seealso{
Other Environmental Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%+must.have\%}()},
\code{\link{\%-=\%}()},
\code{\link{\%missing\%}()},
\code{\link{\%must.have\%}()}
}
\concept{Environmental Integrity}
