% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_integrity.R
\name{\%-=\%}
\alias{\%-=\%}
\title{Remove Objects from an Environment}
\usage{
env \%-=\% x
}
\arguments{
\item{env}{An environment or name of an environment.}

\item{x}{Character vector of object names to remove from \code{env}.}
}
\value{
The target environment, invisibly.
}
\description{
\verb{\%-=\%} is a shorthand wrapper around \code{\link[base:rm]{base::rm()}}.
}
\seealso{
Other Chapter 1 - Environment Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%must.have\%}()},
\code{\link{check.env}()}
}
\concept{Chapter 1 - Environment Integrity}
