% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_environment_integrity.R
\name{\%missing\%}
\alias{\%missing\%}
\title{Missing Environmental Objects}
\usage{
env \%missing\% x
}
\arguments{
\item{env}{(object) An environment or name of an environment}

\item{x}{(string[]) A string vector of the names of objects in query in \code{env}}
}
\value{
A logical vector, the length of \code{x}
}
\description{
\code{\%missing\%} is a wrapper for \code{base::exists} taking the environment or environment name as the \code{LHS} argument and the vector of object names on the \code{RHS}.
}
\seealso{
Other Environmental Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%+must.have\%}()},
\code{\link{\%-=\%}()},
\code{\link{\%check\%}()},
\code{\link{\%must.have\%}()}
}
\concept{Environmental Integrity}
