% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_integrity.R
\name{\%must.have\%}
\alias{\%must.have\%}
\title{Set Required Object Names for an Environment}
\usage{
env \%must.have\% x
}
\arguments{
\item{env}{An environment or name of an environment.}

\item{x}{Character vector of required object names. Use \verb{!!} when passing a
vector or list. Strings beginning with \code{"-"} or \code{"!"} remove required
names.}
}
\value{
Character vector of required object names.
}
\description{
\verb{\%must.have\%} sets an attribute in \code{env} listing object names that the
environment must contain when checked by \code{\link[=check.env]{check.env()}}.
}
\seealso{
Other Chapter 1 - Environment Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%-=\%}()},
\code{\link{check.env}()}
}
\concept{Chapter 1 - Environment Integrity}
