% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_integrity.R
\name{\%+=\%}
\alias{\%+=\%}
\title{Add Objects to an Environment}
\usage{
env \%+=\% x
}
\arguments{
\item{env}{An environment or name of an environment.}

\item{x}{A named list where names are object names and values are object
definitions.}
}
\value{
The target environment, invisibly.
}
\description{
\verb{\%+=\%} is a shorthand wrapper around \code{\link[base:list2env]{base::list2env()}}.
}
\seealso{
Other Chapter 1 - Environment Integrity: 
\code{\link{\%-=\%}()},
\code{\link{\%must.have\%}()},
\code{\link{check.env}()}
}
\concept{Chapter 1 - Environment Integrity}
