% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_environment_integrity.R
\name{\%+must.have\%}
\alias{\%+must.have\%}
\title{Must Have in Addition}
\usage{
env \%+must.have\% x
}
\arguments{
\item{env}{(object) An environment or name of an environment}

\item{x}{(string[]) A string vector of the names that \code{env} must have when checked}
}
\value{
The names of the objects that \code{env} must have in addition to existing required objects, invisibly
}
\description{
\code{\%\+must.have\%} adds additional object names as "must haves" for the given environment
}
\seealso{
Other Environmental Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%-=\%}()},
\code{\link{\%check\%}()},
\code{\link{\%missing\%}()},
\code{\link{\%must.have\%}()}
}
\concept{Environmental Integrity}
