% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_management.R
\name{\%read\%}
\alias{\%read\%}
\title{Snippet Operator}
\usage{
doc \%read\% x
}
\arguments{
\item{doc}{(string) The input source document name given as a string: defaults to the active document when the function is invoked with no argument}

\item{x}{Keywords given as strings or symbols for which a matching snippet is sought. Use \code{c(...)} when passing multiple values.}
}
\value{
see \code{\link{read_snippet}}
}
\description{
This is an infix operator shorthand for \code{\link{read_snippet}} using \code{action=exec}
}
\seealso{
Other Chapter 3 - Workflow Management: 
\code{\link{make_snippet}()},
\code{\link{read_snippet}()},
\code{\link{snippets_toc}()}
}
\concept{Chapter 3 - Workflow Management}
