% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_processing.R
\name{load_unloaded}
\alias{load_unloaded}
\title{Load Unloaded Packages}
\usage{
load_unloaded(
  ...,
  libs = NULL,
  delim = "[,|; ]",
  autoinstall = FALSE,
  chatty = FALSE
)
}
\arguments{
\item{...}{\code{\link[rlang]{dots_list}} A vector of packages to load given as characters.  Delimited strings are allowed: \emph{DO NOT use \code{+} or \code{-}}.}

\item{libs}{(string[]) A vector of packages to load given as characters or symbols.  Delimited strings are allowed: \emph{DO NOT use \code{+} or \code{-}}.}

\item{delim}{(string | "\link{, }") A regular expression delimiter pattern that operates on \code{str}: \emph{DO NOT use glyphs \code{+} or \code{-}}.}

\item{autoinstall}{\itemize{
\item{\code{FALSE} (default): A notification message is provided if \code{chatty} is \code{TRUE}; otherwise, no message is sent}
\item{\code{TRUE}: If a library isn't installed, it will be using the default library installation path}
\item{\code{list()}: Assumes \code{TRUE} and should be a list of arguments (excluding \code{pkgs}) to send to \code{\link[utils]{install.packages}}}
}}

\item{chatty}{(logical | \code{FALSE}) Sets the \code{quietly} argument in the call to \code{\link[base]{library}}()}
}
\value{
See \code{\link[base]{library}}
}
\description{
\code{load.unloaded} checks the packages provided in \code{libs} against a call to \code{\link[base]{search}} and only makes a call to \code{\link[base]{library}} for unloaded (and attached) libraries.
}
\details{
Library names can be declared in a single, delimited string (e.g., \code{"name0 name1, name2|name3"}) or as a vector of strings (e.g. \code{c("name0", "name1", "name2")}).
Inclusions and exclusions can be declared using the following template:\cr  \code{"library_name{+name0+name1+...}"} for inclusions and \code{"library_name{-name0-name1-...}"}.  Since inclusions and exclusions cannot be used in the same call to \code{\link[base]{library}}(), trying to do so with this function will result in an error.
}
\seealso{
Other Chapter 1 - Environment Processing: 
\code{\link{copy_obj}()},
\code{\link{refer.to}()},
\code{\link{save_image}()}
}
\concept{Chapter 1 - Environment Processing}
