% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_workflow_management.R
\name{make.snippet}
\alias{make.snippet}
\title{Make a Snippet}
\usage{
make.snippet(..., include.read = TRUE, use.clipboard = FALSE)
}
\arguments{
\item{...}{Symbols or words serving as keywords that, taken together, distinguish the snippet from others in the same source document}

\item{include.read}{(logical) When \code{TRUE}, an associated \code{\link{read.snippet}} command is provided with argument \code{eval = parse}}

\item{use.clipboard}{(logical) When \code{TRUE}, the snippet contents are saved to the clipboard and \code{`.Last.value`}}
}
\value{
When \code{use.clipboard} is \code{FALSE}, a pair of "tag"-like comments between which code is to be supplied; otherwise, the contents are saved to the clipboard (Windows OS only)
}
\description{
On a new line, \code{make.snippet} creates "tag"-like comments referenced by \code{\link{read.snippet}} (e.g., \code{<snippet: }label\code{>}...\code{<}/snippet\code{>}).  Because of the parsing used, it is important that statements end with a semi-colon (;) as is the case with many other programming languages.  The opening "tag" is created as a code section.
}
\seealso{
Other Code Management: 
\code{\link{read.snippet}()}
}
\concept{Code Management}
