% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_workflow_management.R
\name{make.workflow}
\alias{make.workflow}
\title{Create a Workflow Object}
\usage{
make.workflow(..., wf_name = "new_wf", envir = NULL, eval = TRUE)
}
\arguments{
\item{...}{Unquoted expressions to be executed in the workflow in the order given.  Named arguments are suggested for clarity}

\item{wf_name}{The name of the workflow list object for assignment}

\item{envir}{The target environment for assignment}

\item{eval}{(logical|TRUE) Should the assignment expression be evaluated after creation?}
}
\description{
\code{make.workflow}() assigns an expression list to the target environment.  These objects can then be invoked by calling \code{execute.workflow}()
}
