% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{make_query}
\alias{make_query}
\title{SQL Query Maker}
\usage{
make_query(
  this.conn,
  from,
  from.mod = "",
  where = "1=1 ",
  sel = "*",
  sel.mod = "",
  group_by = NULL,
  having = NULL,
  order_by = NULL,
  ...
)
}
\arguments{
\item{this.conn}{A \code{\link[DBI]{DBI-package}} connection object}

\item{from}{(string) The object segment of a fully-qualified object name}

\item{from.mod}{(string) Table table hints (e.g., \code{WITH (NOLOCK)}) and hard-coded subsequent JOINs}

\item{where}{(string) "WHERE" criteria}

\item{sel}{(string|logical | "*") The columns to select (see 'Details')}

\item{sel.mod}{(string | "") Modifiers to the 'SELECT' statement (e.g., DISTINCT, TOP n, etc ...)}

\item{group_by}{(string) "GROUP BY" criteria}

\item{having}{(string) "HAVING" criteria}

\item{order_by}{(string) "ORDER BY" criteria}

\item{...}{(Not used)}
}
\value{
A (T-)SQL string for use in an RDBMS call
}
\description{
\code{make_query} is a helper function to create a SQL query string with a focus on T-SQL.  For column selection, the following are valid:
\itemize{
  \item \code{"*"} The default, denoting all column names
  \item A vector of column names (e.g., \code{"col1, col2, col3"})
  \item \code{TRUE}: This invokes \code{\link[tcltk]{tk_select.list}} for interactive selection
  \item A string that, when parsed, becomes a single-argument mapper for use in \code{\link[purrr]{keep}} (e.g., \code{"~.x \%ilike\% \"(pat|bp)\""})
  \item A single-argument function serving as the mapper in \code{\link[purrr]{keep}}
}
}
\seealso{
Other Data transmission: 
\code{\link{check.db_conn}()},
\code{\link{export_data}()},
\code{\link{get_data}()}
}
\concept{Data transmission}
