% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_workflow_management.R
\name{mgr_upgrade}
\alias{mgr_upgrade}
\title{Upgrade a workflow Manager}
\usage{
mgr_upgrade(ref)
}
\arguments{
\item{ref}{An unquoted call to a 'workflow_manager' object}
}
\value{
Possibly console messages, but nothing else.
}
\description{
\code{mgr_upgrade} copies predefined class field 'workflows' to a new class instance of the latest (internally-tracked) version.  The contents are returned to an object with the original call defined in \code{ref}
}
