% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_environment_integrity.R
\name{must.have}
\alias{must.have}
\title{Must Have}
\usage{
must.have(env, x)
}
\arguments{
\item{env}{(object) An environment or name of an environment}

\item{x}{(string[]) A string vector of the names that \code{env} must have when checked}
}
\value{
The names of the objects that \code{env} must have
}
\description{
\code{must.have} and related operator \code{\%must.have\%} sets an attribute in the environment given by \code{env} with the name(s) of the object(s) that the environment must have.  Verification is done via \code{\link{env.check}}.
}
\seealso{
Other Environmental Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%+must.have\%}()},
\code{\link{\%-=\%}()},
\code{\link{\%missing\%}()},
\code{\link{env.check}()}
}
\concept{Environmental Integrity}
