% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_management.R
\name{read_snippet}
\alias{read_snippet}
\alias{read.snippet}
\title{Read a Snippet from Source Code}
\usage{
read_snippet(..., doc = NULL, action = "parse", chatty = FALSE)
}
\arguments{
\item{...}{(\code{\link[rlang]{dots_list}}) Keywords given as strings or symbols for which a matching snippet is sought. Use \code{!!!} when passing vectors/lists.}

\item{doc}{The input source document name given as a string: defaults to the active document when the function is invoked with no argument}

\item{action}{(See \code{action} below.)}

\item{chatty}{(logical) Should execution message be generated?}
}
\value{
The snippet text invisibly
}
\description{
\code{read_snippet} returns pre-defined sections (snippets) of larger source files marked with "tag"-like syntax (e.g., \code{<snippet: }label\code{>}...\code{<}/snippet\code{>})
Because of the parsing used, it is important that statements end with a semi-colon (;) as is the case with many other programming languages.
}
\note{
Only the first matching position for the opening snippet tag is used; therefore, ensure that tag labels are unique.
}
\section{\code{action}}{

\itemize{
\item{skip: No code is executed}
\item{parse: Contents of the code region are parsed and printed without execution}
\item{exec: Contents of the code region are parse and executed from the \emph{global} environment}
\item{save: Contents of the code region are saved to the current working directory using the keywords contained in \code{`...`} with file extension \code{'.snippet'}}
\item{html: Contents of the code region are rendered to the viewer using the keywords contained in \code{`...`}}
}
}

\seealso{
Other Chapter 3 - Workflow Management: 
\code{\link{make_snippet}()},
\code{\link{snippets_toc}()}
}
\concept{Chapter 3 - Workflow Management}
