% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_processing.R
\name{refer.to}
\alias{refer.to}
\title{Refer to an Environment}
\usage{
refer.to(x)
}
\arguments{
\item{x}{An environment or character string}
}
\value{
Invisibly, the environment if it exists or an error message
}
\description{
\code{refer.to} is a convenience function for referring to an environment. It is a wrapper for \code{\link{as.environment}} that allows for the use of character strings and environments as arguments.
}
\seealso{
Other Chapter 1 - Environment Processing: 
\code{\link{copy_obj}()},
\code{\link{load_unloaded}()},
\code{\link{save_image}()}
}
\concept{Chapter 1 - Environment Processing}
