% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{terminate.cluster_workers}
\alias{terminate.cluster_workers}
\title{Terminate Remote Cluster Workers}
\usage{
terminate.cluster_workers(cl, chatty = FALSE)
}
\arguments{
\item{cl}{A socket cluster object make with package 'parallelly' (e.g., \code{\link[parallelly]{makeClusterPSOCK}}) or similar}

\item{chatty}{(logical | FALSE) Verbosity flag}
}
\value{
None
}
\description{
Terminate Remote Cluster Workers
}
\section{Required}{

The following \href{https://docs.microsoft.com/en-us/sysinternals/downloads/pstools}{PSTools} programs must be globally accessible on the active session workstation: \cr \itemize{\item{psexec} \item{pskill} \item{tasklist}}
}

\seealso{
Other Parallelism: 
\code{\link{do.make_cluster}()},
\code{\link{get.cluster_meta}()}
}
\concept{Parallelism}
