% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_book.of.workflow.R
\name{book.of.workflow}
\alias{book.of.workflow}
\title{Book of Workflow Overview}
\description{
The following functional families are covered in \code{book.of.workflow}:
}
\section{Chapter 1 - Environment-processing}{

The focus is on loading libraries and managing environment objects:
\itemize{
	\item{\code{\link{load_unloaded}}}
	\item{\code{\link{save_image}}}
	\item{\code{\link{copy_obj}}}
}
}

\section{Chapter 2 - Environment Integrity}{

 This family of functions is focused on maintaining the integrity of object environments, primarily, to mitigate errors due to  missing object dependencies:
 \itemize{
 	\item{\code{\link{check.env}}}
 	\item{\code{\link{\%must.have\%}}}
 	\item{\code{\link{\%+=\%}}}
 	\item{\code{\link{\%-=\%}}}
 }
}

\section{Chapter 3 - Workflow Management}{

 This family of functions focuses on code execution workflow:
 \itemize{
 	\item{\code{\link{read.snippet}}}
 	\item{\code{\link{make.snippet}}}
 	\item{\code{\link{snippets_toc}}}
 }
}

