% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_integrity.R
\name{check.env}
\alias{check.env}
\title{Environment Integrity Check}
\usage{
check.env(...)
}
\arguments{
\item{...}{One or more environments to check}
}
\value{
If all objects that \code{env} \code{\link{\%must.have\%}} are present, nothing is returned; otherwise, an error message is returned.
}
\description{
\code{check.env} checks one or more environments for required objects
}
\seealso{
Other Chapter 2 - Environment Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%-=\%}()},
\code{\link{\%must.have\%}()}
}
\concept{Chapter 2 - Environment Integrity}
