% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_environment_processing.R
\name{do.get_pkgs}
\alias{do.get_pkgs}
\title{Install and load packages}
\usage{
do.get_pkgs(pkg.list, is.custom = FALSE, custom.root = "")
}
\arguments{
\item{pkg.list}{(string[]) A string vector of package names or a delimited string ("[,|; ]")}

\item{is.custom}{(logical | \code{FALSE}) When \code{TRUE}, the value of \code{custom.root} is used under a "local source" context}
}
\value{
Returns the success or failure of each package (down)load
}
\description{
\code{do.get_pkgs} Serves as a wrapper for \code{\link[base]{install.packages}}.
}
