% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_environment_integrity.R
\name{\%-=\%}
\alias{\%-=\%}
\title{Remove Objects from an Environment}
\usage{
env \%-=\% x
}
\arguments{
\item{env}{(environment, string) An environment or name of an environment}

\item{x}{(string[]) A collection of strings representing the names of the objects to remove from \code{env}}
}
\value{
The target environment, invisibly
}
\description{
Operator \code{\%-=\%} is a wrapper for \code{base::rm()}
}
\seealso{
Other Environmental Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%+must.have\%}()},
\code{\link{\%missing\%}()},
\code{\link{env.check}()},
\code{\link{must.have}()}
}
\concept{Environmental Integrity}
