% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_integrity.R
\name{\%must.have\%}
\alias{\%must.have\%}
\title{Must Have}
\usage{
env \%must.have\% x
}
\arguments{
\item{env}{(object) An environment or name of an environment}

\item{x}{(string[]) A vector or strings containing the object names that \code{env} must have when checked.  Use the \code{\link[rlang]{!!}} operator when passing a vector or list.  Each string beginning with "-" or "!" will remove object names the target environment must have.}
}
\value{
The names of the objects that \code{env} must have
}
\description{
\code{\%must.have\%} sets an attribute in the environment given by \code{env} with the name(s) of the object(s) that the environment must have.  Verification is done via \code{\link{check.env}}.
}
\seealso{
Other Chapter 2 - Environment Integrity: 
\code{\link{\%+=\%}()},
\code{\link{\%-=\%}()},
\code{\link{check.env}()}
}
\concept{Chapter 2 - Environment Integrity}
