% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_management.R
\name{read_snippet}
\alias{read_snippet}
\title{Read a Snippet from Source Code}
\usage{
read_snippet(..., doc, action)
}
\arguments{
\item{...}{(\code{\link[rlang]{dots_list}}) Keywords given as strings or symbols for which a matching snippet is sought. Use \code{!!!} when passing vectors/lists.}

\item{doc}{The input source document name given as a string: defaults to the active document when the function is invoked with no argument}

\item{action}{One of \code{goto}, \code{skip}, \code{parse}, \code{exec}, \code{save}}
}
\value{
The snippet text invisibly
}
\description{
\code{read.snippet} returns pre-defined sections (snippets) of larger source files marked with "tag"-like syntax (e.g., \code{<snippet: }label\code{>}...\code{<}/snippet\code{>})
Because of the parsing used, it is important that statements end with a semi-colon (;) as is the case with many other programming languages.
}
\details{
With the exception of \code{action = goto}, the document cursor moves to the closing snippet tag (i.e., \code{"</snippet>"}) when using RStudio.
}
\section{\code{action}}{

\itemize{
\item{goto (RStudio-only): Editor cursor jumps to beginning of code region without execution}
\item{skip (RStudio-only): Editor jumps to end of code region without execution}
\item{parse: Contents of the code region are run through \code{\link[base]{cat}} without execution}
\item{exec: Contents of the code region are parse and executed from the global environment}
\item{save: Contents of the code region are saved to the current working directory using the keywords contained in \code{`...`} ending in '.snippet'}
}
}

\seealso{
Other Chapter 3 - Workflow Management: 
\code{\link{make_snippet}()},
\code{\link{snippets_toc}()}
}
\concept{Chapter 3 - Workflow Management}
