% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_workflow_management.R
\name{snippets_toc}
\alias{snippets_toc}
\title{Snippets Table of Contents}
\usage{
snippets_toc(doc)
}
\arguments{
\item{doc}{The path to a document containing code sections created via \code{\link{make.snippet}}()}
}
\value{
Invisibly, a listing of snippet code sections for the document provided
}
\description{
\code{snippets_toc} creates a table of contents of snippet code sections.
}
\note{
An active session must be required to use this file if \code{doc} is \code{NULL}
}
\seealso{
Other Chapter 3 - Workflow Management: 
\code{\link{make.snippet}()},
\code{\link{read.snippet}()}
}
\concept{Chapter 3 - Workflow Management}
