% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_external_data_management.R
\name{terminate.cluster_workers}
\alias{terminate.cluster_workers}
\title{Terminate Remote Cluster Workers}
\usage{
terminate.cluster_workers(...)
}
\description{
`terminate.cluster_workers()` is no longer in use
}
\seealso{
Other DEPRECATED: 
\code{\link{do.get_pkgs}()},
\code{\link{do.make_cluster}()}
}
\concept{DEPRECATED}
